rework for nft test passing

+++ x/nft/genesis.go
@@ -16,7 +16,7 @@ func InitGenesis(ctx sdk.Context, k keeper.Keeper, genState types.GenesisState) 
 	}
 
 	// Set nft count
-	k.SetNftCount(ctx, uint64(len(genState.NftList)))
+	k.SetNftCount(ctx, uint32(len(genState.NftList)))
 
 	// this line is used by starport scaffolding # ibc/genesis/init
 }
